Move RobotSimulatorChangeDynamicsArgs struct to its own header
